drivers: adc: permit use of vref-mv, even if reference == ADC_REF_INTERNAL
It is useful to be able to provide a `zephyr,vref-mv` property that
takes a potential divider into consideration, allowing direct
measurement of a voltage higher than the part's maximum, with no further
conversion steps.

For example, a potential divider with a 10kOhm resistor between the
signal and the ADC input, and a 1kOhm resistor between the ADC input and
ground can permit measuring a voltage up to 11x higher. In this case, a
`zephyr,vref-mv` value of `Vref * 11` can be given. When a call is made
to `adc_raw_to_millivolts_dt()`, the "real" voltage is returned
immediately.

Signed-off-by: Attie Grande <attie.grande@argentum-systems.co.uk>